+++ src/OpenTelemetry.Instrumentation.Runtime/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Updated OTel SDK package version to 1.3.0
   ([#411](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/411))
+* Fix some bugs in Runtime metrics
+  ([#409](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/409))
 
 ## 0.1.0-alpha.1